requirements.txt
django
django-environ
pillow
stripe
django-allauth
django-tailwind
django-crispy-forms
crispy-tailwind
psycopg2
djangorestframework



